refactor: unify transaction display modes and decoding states
