+++ packages/ui/src/DecodedTransactions/TransactionDisplay/TransactionDisplay.tsx
@@ -1,4 +1,5 @@
 import { DecodedTransaction } from '@buildeross/hooks/useDecodedTransactions'
+import { useTransactionSummary } from '@buildeross/hooks/useTransactionSummary'
 import {
   CHAIN_ID,
   DaoContractAddresses,
@@ -6,10 +7,21 @@ import {
   ProposalTransactionBundleContext,
   SimulationOutput,
 } from '@buildeross/types'
+import { Stack } from '@buildeross/zord'
 import React from 'react'
 
-import { DecodedDisplay } from './DecodedDisplay'
-import { NonDecodedDisplay } from './NonDecodedDisplay'
+import { ArgumentDisplay } from '../ArgumentDisplay'
+import { TransactionAiSummary } from './TransactionAiSummary'
+import {
+  DecodedCallDisplay,
+  DisplayModeToggle,
+  RawCalldataDisplay,
+  SimulationWarning,
+  TransactionDisplayShell,
+} from './TransactionDisplayShared'
+import { useDecodedTransactionContext } from './useDecodedTransactionContext'
+
+const DISABLE_AI_SUMMARY = process.env.NEXT_PUBLIC_DISABLE_AI_SUMMARY === 'true'
 
 export const TransactionDisplay: React.FC<
   DecodedTransaction & {
@@ -19,6 +31,7 @@ export const TransactionDisplay: React.FC<
     proposalMetadata?: ProposalDescriptionMetadataV1
     bundleContext?: ProposalTransactionBundleContext
     simulation?: SimulationOutput
+    isDecoding?: boolean
   }
 > = ({
   chainId,
@@ -27,40 +40,138 @@ export const TransactionDisplay: React.FC<
   proposalMetadata,
   bundleContext,
   simulation,
+  isDecoding = false,
   ...decoded
 }) => {
-  if (decoded.isNotDecoded || !decoded.transaction.args) {
-    const calldata = decoded.isNotDecoded
-      ? decoded.transaction
-      : decoded.transaction.encodedData
+  const [showRawCalldata, setShowRawCalldata] = React.useState(false)
 
-    return (
-      <NonDecodedDisplay
-        {...{
-          chainId,
-          target: decoded.target,
-          calldata,
-          value: decoded.value,
-          simulation,
-          index,
-        }}
-      />
-    )
-  }
+  const isDecoded = !decoded.isNotDecoded && !!decoded.transaction.args
+
+  const {
+    sortedArgs,
+    escrowData,
+    streamData,
+    tokenMetadata,
+    nftMetadata,
+    isLoadingMetadata,
+    isSendWithValue,
+  } = useDecodedTransactionContext({
+    chainId,
+    transaction: isDecoded
+      ? decoded.transaction
+      : {
+          functionName: 'send',
+          args: { value: { name: 'value', value: decoded.value, type: 'uint256' } },
+          functionSig: '',
+          encodedData: '0x',
+          argOrder: ['value'],
+        },
+    target: decoded.target,
+    value: decoded.value,
+    enabled: isDecoded,
+  })
+
+  const transactionData = React.useMemo(() => {
+    if (!isDecoded || isLoadingMetadata || DISABLE_AI_SUMMARY) return null
+
+    return {
+      chainId,
+      addresses,
+      transaction: {
+        args: decoded.transaction.args,
+        functionName: decoded.transaction.functionName,
+      },
+      target: decoded.target as `0x${string}`,
+      tokenMetadata: tokenMetadata || undefined,
+      nftMetadata: nftMetadata || undefined,
+      escrowData: escrowData || undefined,
+      streamData: streamData || undefined,
+      proposalMetadata,
+      bundleContext,
+    }
+  }, [
+    addresses,
+    bundleContext,
+    chainId,
+    decoded,
+    escrowData,
+    isDecoded,
+    isLoadingMetadata,
+    nftMetadata,
+    proposalMetadata,
+    streamData,
+    tokenMetadata,
+  ])
+
+  const {
+    summary: aiSummary,
+    error: errorSummary,
+    isLoading: isGeneratingSummary,
+    mutate: regenerateSummary,
+  } = useTransactionSummary(transactionData)
+
+  const rawCalldata = decoded.isNotDecoded
+    ? decoded.transaction
+    : decoded.transaction.encodedData || 'Unavailable'
+
+  const toggleMode: 'toggle' | 'decoding' | 'rawOnly' = isDecoded
+    ? 'toggle'
+    : isDecoding
+      ? 'decoding'
+      : 'rawOnly'
 
   return (
-    <DecodedDisplay
-      {...{
-        index,
-        chainId,
-        addresses,
-        proposalMetadata,
-        bundleContext,
-        simulation,
-        transaction: decoded.transaction,
-        target: decoded.target,
-        value: decoded.value,
-      }}
-    />
+    <Stack style={{ wordBreak: 'break-word', overflowWrap: 'break-word' }} w="100%">
+      <DisplayModeToggle
+        mode={toggleMode}
+        showRawCalldata={showRawCalldata}
+        onToggle={() => setShowRawCalldata((state) => !state)}
+      />
+      <TransactionDisplayShell chainId={chainId} target={decoded.target} index={index}>
+        {isDecoded && !showRawCalldata ? (
+          <DecodedCallDisplay
+            functionName={decoded.transaction.functionName}
+            value={decoded.value}
+            sortedArgs={sortedArgs}
+            isSendWithValue={isSendWithValue}
+            renderArg={(argKey, i) => {
+              const arg = decoded.transaction.args[argKey]
+
+              return (
+                <ArgumentDisplay
+                  chainId={chainId}
+                  key={`${argKey}-${arg.name}-${i}`}
+                  arg={arg}
+                  target={decoded.target}
+                  functionName={decoded.transaction.functionName}
+                  tokenMetadata={tokenMetadata}
+                  nftMetadata={nftMetadata}
+                  escrowData={escrowData}
+                  streamData={streamData}
+                />
+              )
+            }}
+          />
+        ) : (
+          <RawCalldataDisplay calldata={rawCalldata} value={decoded.value} />
+        )}
+      </TransactionDisplayShell>
+
+      <SimulationWarning simulation={simulation} />
+
+      <TransactionAiSummary
+        isVisible={
+          isDecoded &&
+          !isLoadingMetadata &&
+          !DISABLE_AI_SUMMARY &&
+          !errorSummary &&
+          !isGeneratingSummary
+        }
+        isGeneratingSummary={isGeneratingSummary}
+        aiSummary={aiSummary}
+        errorSummary={errorSummary}
+        onRegenerate={() => regenerateSummary()}
+      />
+    </Stack>
   )
 }
